Termbox: Mark mul descriptions as not applicable
For accessibility, this will add a aria-label to
the (disabled) description textarea which screen
readers will read instead of the (shorter) placeholder
message.

This was tested with TalkBack on Android 11 with
both Firefox and Chrome. While in Chrome this works
flawlessly, Firefox will, after reading the accessibility
label, still announce that text can be entered in the
disabled input (probably due to
https://bugzilla.mozilla.org/show_bug.cgi?id=1356018)

Bug: T338302
Change-Id: Iae14314a404299555d8fd2754149f76055d440b4