REST: Handle conditional requests for GET property
This modifies the PreconditionMiddleware, so that it is capable of
dealing not only with item revision metadata, but generally that of
any entity revision.

The end-to-end test of conditional requests is modified in such a way,
that the new GetProperty route is added to the set of routes to be
tested and the corresponding test property metadata will be used for it,
instead of the test item metadata.

Bug: T338138
Change-Id: I144ff7a0249d8ca81f0879e6f4ab845da617592d