SECURITY: Escape quotes in js templates
Previously, the code in wbTemplate would not escape quotes in
parameters, which resulted in an XSS-vulnerability when those parameters
when then used in attributes of HTML tags, for example the `title=""`
attribute.

This basically undoes the production code changes made in I641725a254f
for T44956. However, it does not change the tests added in that change,
so I wonder if "HTML-encoded non-breaking-space causes invalid HTML" was
really the root issue that was going on. In any case, on my machine, the
Item page in French seems to work perfectly fine, including editing.

Bug: T339111
Change-Id: Ib2a85fc6f1d1339fb857ee749d3a49ac19aece36