Decouple Template from Message
As far as I can tell, the only Message functionality we’re using is
parameter replacement ($1, $2, etc.) – we don’t process variables or
magic words, and we don’t parse wikitext. Let’s simplify this process
and decouple the functionality from the big Message class.

In the longer term, we should replace these custom templates with
MediaWiki’s standard solution based on Mustache (which is available in
both PHP and JS – we need both). However, that’ll be more work because
we’ll have to update the templates and all template calls, to replace
numeric variables like $1 with named variables like {{whatever}}.

Motivated by T339111.

Bug: T91067
Change-Id: If8cdcfac15bbfe158ef2cb08187c3cbcd751a853