SetClaimValueTest: Use string properties
This test uses commonsMedia properties since its itroduction in 2012,
in b2c768953d12. Adding commonsMedia Snaks invokes secondary data
updates, that (try to) access the media storage, which is undesirable
in tests. Given that these side effects/ secondary data updates
didn't exist when the original change was made, it is safe to assume
that this was not done in an attempt to also test these.

Bug: T340018
Change-Id: I423517b480d1eea5b8068b3442f4450f352e9440