Mark `mul` descriptions as N/A in legacy Termbox via js
`mul` Multiple Languages cannot have descriptions and that should be
indicated to the user viewing an Entity, both visually as well as with
screen readers. The first part of this was done in I5ddd3c5fb
(commit a119f4e).

Given that this only affects descriptions, and only for one specific
language code, and only in one place in the file, the logic has been
added completely to descriptionview.js, so that
entitytermsforlanguagelistview.js doesn't need to know about it.

Also, this does not try to set the `lang="..."` attribute to the
language that the message in the `title="..."` is actually in, because:
1) this does plain not seem possible: mw.message does not seem to have
   that info
and
2) this is also actually not implemented in PHP. Even though we're
   calling LocalizedTextProvider::getLanguageOf( $key ) there, that key
   is actually ignored in the implementation MediaWikiLocalizedTextProvider

Bug: T329644
Change-Id: If687dfe5f9a541ee3c146cc4f4a189231f341504