Rename Mediawiki* classes to MediaWiki*
This has bothered me for years. MediaWiki is spelled with an uppercase
W, and this is important because at least IntelliJ (and probably other
IDEs) let you search for “abbreviated” class names, but will only use
uppercase letters as “word breaks”, i.e. if the class is named
MediawikiEditEntity, you can search for it as “MedEdEnt”, but not
“MWEdEnt”. Fix this.

Note that class names are case-insensitive in PHP, so this is not a
breaking change, and other extensions that reference these classes
(WikibaseLexeme and Nimiarkisto) will continue to work whether they
update their references to these classes or not.

Change-Id: Ia9ac1f43822b5750fb35835ad0a1f830905a6cc1