Add MUL as last language in Termbox if any MUL term exists
This is the client side part only, the server side
termbox rendering will be changed in I9d214fdb21c.

This is going to result in more mismatches with the
server generated DOM: We want to have "mul" in the
last position always (unless the user has "mul" in
their babels), which is the same behavoir we expect
in the server rendering. But, for logged out users,
we might dynamically add more languages from ULS,
meaning we need to discard the previous "mul" entry
and re-create it (but if adding languages from ULS,
we we were already touching the DOM anyway).

Bug: T339103
Change-Id: I1778ea90ebbf89bb6083e3fe324219878f3c9033