OutputPageBeforeHTMLHookHandler: Log when loading full entities
This also very slightly changes the behaviour of
OutputPageBeforeHTMLHookHandler::needsRealEntity
because an empty array (instead of null) should not
result in loading the full entity.
While OutputPageBeforeHTMLHookHandler::logger can
theoretically be a NullLogger, it will always be
a real instance in production code (set via the
factory method), as far as I can tell.

Bug: T340642
Change-Id: Id9dff338709946d33c196ebc40acf53cd67cd0c0