Add explicit `global` in WikibaseClient.ci.php
When this file is not included from the global scope, as is the case
with PHPUnit's bootstrap in I95a80c44f8d88c3c498efe1ae64008f0ff1eea55,
the variable is only set locally. The bootstrap code has some hacks to
export all variables to the global scope, but that only runs after the
whole Setup.php has been loaded.

ExtensionLoadHandler, which runs after these settings are loaded but
before the end of Setup.php, reads these values from the global state,
so we need to make sure that the values are already global by then.

Bug: T90875
Change-Id: I97b1e10aac30213574844166cb3925affc3b4c5c