Add feature flag to always show MUL in Termbox (PHP)
We want to encourage the usage of the `mul` language code in Labels and
aliases for Entities, so it should always be shown in the Termbox.

In order to introduce this slowly to the community, another
feature flag is introduced specifically for this "show always"
functionality. This will allow us to do a "soft rollout". That feature
flag will eventually be removed in T330217.

This change affects both Items as well as Properties and all future
Wikibase Entities that support Labels or Aliases.

Bug: T339104
Change-Id: If9cb0b8d3b7a9b0bc8d84b9cfe4cd4373791561b