Always show `mul` in the desktop Termbox, depending on the feature flag
This is JS-leg to the change in If9cb0b8d3b7a9.

This change is optimized for the temporary feature flag and its
associated code to be easily removed.

The alternative to doing this change in entitytermsforlanguagelistview
would have been to do it where the user languages are first assembled,
in `wikibase.getUserLanguages()`.

Bug: T339104
Depends-On: I310948c0875548b2b2fb7954c5656e78d02d83df
Change-Id: I715801620660f66eda5eb30140386c47389250dc