Set allowDataAccessInUserLanguage in UsageTrackingIntegrationTest
The test only works if this setting is true. Also, since we moved the
settings (and everything else) to the service container a while ago, we
can make the way we change the settings a bit nicer: clone them, modify
what we need, override the service, and let MediaWikiIntegrationTestCase
reset the service at the end.

Bug: T340928
Change-Id: Ie084c56b5e08d373e666731c50182b62c89dfb84