Fix styles for js-generated mul description placeholder
table row for `mul`, because other languages where added by uls (see
I1778ea90eb, commit 5bafe18c5), it re-creates it with the default
descriptionview template, instead of using the mul-specific one. For
that template, the styles added in change I5ddd3c5f (commit a119f4e) no
longer match.

To fix that, this commit adjusts the styles so that they match both templates.
Also, it consolidates the styles to make use of the nesting-capabilities
of less.

Bug: T329644
Change-Id: Icfe96f0f111908fa64793de502b4c61f31b9e1b9