OutputPageBeforeHTMLHookHandler: Don’t load full entities
The log message added in change Id9dff33870 (commit f3e2059bff) has
never been seen in production, so it seems like there’s no reason to
expect the parser output / output page property to ever be missing.
Let’s remove the fallback code to load the entity from the entity
lookup, and throw an exception if the property is missing.

This makes the handler’s EntityRevisionLookup unused, but I want to keep
it for a bit – I think it’s possible (though not certain) that we’ll
need it for T340832 later.

Bug: T340642
Change-Id: I4785304574cf7b927b4f52522a30051041e71e96