Use fallback chain for label placeholders in legacy termbox
To disincentivize users from adding redundant labels, we want to show
fallbacks for missing labels, excluding an implicit 'en' fallback where
applicable.

This commit in particular adjusts the rows of the legacy Termbox that
are rendered into the parser cache. Followup commits will adjust the
behavior of user-specific server side rendered rows (that is,
babel-based rows that do not contain any Terms, I7901aaf6), and
client-side rendered rows (that is, rows from ULS and the ones in "All
Languages", I61f7fea0).

In order to not trigger larger-scale adjustments, this determines
whether `en` is an implicit fallback based on whether it appears at the
end of the fallback chain, that is, after `mul`. However, if this
feature is ever disconnected from `mul` being enabled or not, then this
information will need to be properly integrated into the fallback chain
itself.

This change now includes the structure from I816178e3a21.

Overall, this feature is currently intended to be an experiment, run on
test.wikidata.org.

Co-Authored-By: Lucas Werkmeister <lucas.werkmeister@wikimedia.de>
Bug: T340832
Change-Id: I132a8518eb8d17a9b7c09f01aed092895def17bd