Set entity labels in ParserOutput to show as fallback in termbox
This change enables us to also show the label-placeholder fallbacks for
rows that are coming from babel and that have no Term at all in them
yet. Those are not part of the parser output, but are created new on
every page view.

To be able to still make the fallback happen without loading the entity
from the database, we store the existing labels in a separate parser
cache key.
In terms of additional caching amount used, this should be
much less than the existing stored HTML for every row with Terms.
In terms of additional processing per Entity page view, this is
hopefully also manageable as those rows would be created anyway, now
just a small amount for accessing the fallback chain is added.

The code adding the data to the parser cache was moved in I99c16a551
(commit 52cf0c1). So this change also adjusts all the comments to match.

Bug: T340832
Change-Id: I7901aaf625fe3ea1309d29b7c02ac134ee1ea74f