Refactor TermsListView template calls
Only render each template once (unify the two render calls in
getAliasesView()), and pull as much logic as makes sense out of the
render call (replace a bunch of ternary expressions). This should help
keep the code relatively clean as we add more logic for T340832, as well
as maybe help with porting these templates to Mustache later (T91067).

Change-Id: Ibb4950d7ac59596c838aeccb447fbaec88fc4acc