Termbox: Initialize using new init function
This changes the way Termbox is initialized (see ADR 23) using the new
function exported from wikibase.termbox.init.js. It keeps the
functionality of the existing `wikibase.termbox` ResourceLoader module
as it was and adds a new `wikibase.termbox.init` module which exports
the init function to be used by Termbox consumers.

Bug: T342241
Change-Id: Icbf3ace57643fac1a8942acdb33ee5578862aa65