Check for false repoDomain explicitly in RepoDomainDbFactory
The domain ID can be the empty string when the DB storage is disabled
(via LoadBalancerDisabled). This is not something that can generally
occur, except for the installer and no-database maintenance script.
Another scenario in which this will occur is PHPUnit tests not in the
database group, see I96ecf9ff.

The original intent of this check was to avoid callers passing boolean
false (which has a special meaning), so remove the typehint and use
Assert to verify that the parameter is a string.

Change-Id: I5a529b3322ce63da4dba880261a3290d796716f8