Show fallback labels as placeholders while editing
We want to give editors a consistent experience when seeing the
fallback labels for languages with out a dedicated Label in the legacy
termbox both when viewing an Item/Property and when editing it.

For that, this change reuses the method to calculate the placeholder
label from I61f7fea0. That way we should be able to ensure that they're
always consistent.

Bug: T338330
Change-Id: I435586f063011cdd1822014a316472e660008e52