Inject LanguageNameLookupFactory into WikibaseValueFormatterBuilders
Instead of injecting a LanguageNameLookup in the user language (which we
really shouldn’t be using in the service container), inject a factory,
then get the right language from the FormatterOptions when we actually
need a specific lookup. This also means that the language names are now
consistent with the rest of the formatted value if the language option
is set to something other than the request user language.

The logic in newLanguageNameLookup() – using either OPT_LANG or
OPT_LANGUAGE_FALLBACK_CHAIN – is inspired by similar code in
FormatterLabelDescriptionLookupFactory::getLabelDescriptionLookup(),
though here I think it makes more sense to use OPT_LANG first, because
we wouldn’t use anything but the first element of the language fallback
chain anyways.

newMonolingualFormatter() needs an $options parameter added, but it’s
not called outside of Wikibase, so that should be fine. (This also makes
it more consistent with the other methods anyway.)

If the FormattingOptions are missing a language option, log a warning
and fall back to $wgLang; assuming the warning doesn’t show up in
production, we can later change this to throw an exception instead.
Unfortunately, I don’t think we can usefully log the $options;
FormatterOptions doesn’t offer a way to list the defined options / keys,
and putting the whole $options object into the log context doesn’t end
up showing the inner options either – I tested this at [1], where
non-empty options just got logged as

    {
      "ValueFormatters\\FormatterOptions": []
    }

[1]: https://logstash.wikimedia.org/app/discover#/doc/logstash-*/logstash-mediawiki-1-7.0.0-1-2023.08.01?id=VZ6FsIkBuj8l-JjZS1ip

Bug: T281726
Change-Id: I6db4b0332c99694ddadef9298ddf0472d509079b
(cherry picked from commit 30a40a791e4c0d3ed7198b52c065be29b4eb1b0e)