Replace use of ParserOutput::addJsConfigVars(), deprecated since 1.38
Most uses can be replaced with ::setJsConfigVar(); the major semantic
change is that ::setJsConfigVar() is not expected to overwrite
previous values, but that doesn't appear to be an issue in most uses
here.

ParserOutputJsConfigBuilder was slightly refactored to take the
$parserOutput as an argument, more suited to the desired write-only
behavior of ParserOutput as a collector of metadata.

GlobeCoordinateKartographerDataUpdater was refactored to use the
ParserOutput::collectMetadata() method (it could equivalently use one
of the other ParserOutput merge methods) so that it is less tightly
coupled to the exact properties which Kartographer is setting.

Bug: T300307
Bug: T305161
Change-Id: I39ae7d7a40190eedaa024097a6442cd02b6a02e7