Create full page in EntityIdLocalPartPageTableEntityQueryDbTest
The test is implicitly relying on localPartOne having a revision with ID
1 (page_latest). This is currently true because addCoreDBData always
creates a test page, but that's going to be removed soon.
Create a full test page instead, using getExistingTestPage because it's
much simpler to read, and because we don't need the page to have
anything special.

Also capitalize the page title, or WikiPageFactory would fail because a
lowercase title is considered invalid.

Bug: T342428
Change-Id: Ic64ded5e2c0b59e7c888ece9566076058a125be4