Avoid DB access in REST handler tests
None of these tests is in the database group, so it should not access
the database, not even indirectly. All the violations are caused by
EntityRevisionLookup, so replace it with a mock that doesn't find
anything (which is the expected behaviour if the DB is not needed).

Create a trait for the mocking method, because it's used a lot; move
there the one for ChangeTagsStore, too.

Follow-up to, and partial revert of, Icb3b3ad7, due to the new tests
added in I3adfcede.

Change-Id: Idab06dc9838e8254da7066d6f8a6c38cec932b9a