Use InsertQueryBuilder
In DatabaseTermStoreWriterBase and SqlSubscriptionManager, we need to
make sure to skip the call if the rows are empty – InsertQueryBuilder,
unlike IDatabase::insert(), doesn’t tolerate empty $rows. All the other
places that use ->rows() (rather than ->row()) already don’t call
->insert() if the rows are empty, as far as I can tell (either due to an
explicit check near the ->insert() call or a bit further up the call
stack, or due to how the rows or batches are assembled).

Skipping the tests for now, because there’s a lot more calls in there.

Change-Id: I596f3d34927fcb8b53112a1c37fef99d871d815d