tests: Use addDBDataOnce instead of DIY implementation
MediaWikiIntegrationTestCase provides the addDBDataOnce method, whose
function is to make DB changes that are not cleared until the whole test
class has finished. This is going to be enforced in T342301.

The current approach (using a static property) does not tell the parent
class that the data should remain after each tests, so after T342301
tests would fail.

A downside of using addDBDataOnce is that it runs before setUp, so some
fixtures need to be created on-demand. Unfortunately, there isn't much
we can do about this.

Also, in BotEditTest, test sets 8 and 9 would fail when run on their own
because they interfere with previous test sets. Use different titles and
update the data provider to reflect the correct values.

Bug: T342301
Change-Id: I780e4969b6d80c49b3eab71b82d1eae8f9c16f90