tests: don't set client service if client is not enabled
This broke our secondary CI job that runs with only repo enabled. It
triggered the following error message:

Wikimedia\Services\NoSuchServiceException: No such service: WikibaseClient.SiteGroup

If client is enabled, then that service is accessed in the
testGetParserOutput and testPageProperties tests because they trigger
the ContentAlterParserOutput hook to which client is subscribed. That
service is then accessed when injecting the dependencies for that
handler.

This maybe could use some more overall untangling, but for now this
change should hopefully make our secondary CI green again.

Bug: T344416
Change-Id: I600d76250b639fe1d95ad028a30c4594079d3dc6