Make certain PHPUnit test setups more robust
I ended rewriting EntitySourceDocumentUrlProviderTest to turn it into a
pure unit test that doesn't use any live services. I think it was meant
to be like this from the beginning. There is not much to assert in the
class under test. The old test was actually testing features from
core's TitleFactory. I don't think this is needed to prove that and how
EntitySourceDocumentUrlProvider works.

This is motivated by Ia408c89.

Change-Id: I144145d76a24f11bbc37d4652ad504f46f05508e