SECURITY: Add rate limits and edit filters to item merging
EntityRedirectCreationInteractor was missing rate limiting, and
ItemMergeInteractor was missing that plus edit filter running (e.g.
AbuseFilter). Add both.

The way that errors are reported isn’t great, but I don’t want to do the
larger refactoring that would be required to do it better in a security
patch. (But once this is published on Gerrit, I think I’d like to change
these classes to return a Status instead of throwing exceptions.)

Bug: T345064
Change-Id: Ia2a4a248f891282f2aa03afd34f4643b517b8ef7