Fix tainted references dependency conflicts
This also deals with some of the eslint warnings and errors that came up
and disables some others, to be tackled later.

The new version of "eslint-config-wikimedia" requires several new
plugins. When updating the dependencies of this package properly,
it should be evaluated if those packages and the (upstream) eslint
rules that require them are actually desirable in this context here.

Bug: T345725
Bug: T345660
Change-Id: I0b89c42261c3ba1e4db21c3305c8e88671f30982