Add anon edit warning link parameters
The previous parameter was not used in any translation of that message
anymore, so replacing it should not be an issue at all.

Sadly, the inner parts of the AnonymousEditWarningBuilder rely on so
many large classes that unit test is not meaningful. An integration
tests is used instead and asserts the parts that do not vary with the
local setup.

The SpecialPageFactory can be injected into the Action once that has
been refactored to using dependency injection in
I29941ad09ce48d655c35f2c64b7d23a637061998.

Bug: T330550
Change-Id: Ic102157019f6ee9852fa1fcb8627a86176af5f1c