EntityId: Soft-deprecate Serializable methods and stop using them
Most callers of serialize() should use getSerialization() instead (tests
for the method specifically are left alone for now); unserialize()
doesn’t even have any callers outside of its tests and __unserialize()
ipmlementations.

While it was never guaranteed that serialize() and getSerialization()
return the same string, this is indeed the case for all EntityId classes
I could find, and a lot of the callers were relying on it anyway.

Bug: T345856
Change-Id: I64e2e54b344e8dd64078f81502d0d067f6ea3a5b