Inject EntitySearchHelper into SpecialItemDisambiguation
I don’t think there was a reason to inject EntitySearchHelperCallbacks
and construct the TypeDispatchingEntitySearchHelper manually – this code
seems to predate the service migration, so it probably just never came
up until now.

In theory, this now includes some Federated Properties code if fed props
are enabled, but FedPropertiesTypeDispatchingEntitySearchHelper just
forwards to the regular TypeDispatchingEntitySearchHelper if the entity
type isn’t “property”, and SpecialItemDisambiguation (naturally) only
searches for entity type “item”, so that should be fine.

Change-Id: I7b2ca2234c103b162cb9a5ea17cd1bb42007c016