Improve SpecialItemDisambiguation service injection
Replace the deprecated LanguageNameLookup service with a
LanguageNameLookupFactory that we can call with our context language
later, and introduce a new ItemDisambiguationFactory from which we can
get an ItemDisambiguation later but which the tests can mock to return a
mock ItemDisambiguation instead.

See also Id96a3a8fa5i on Gerrit for an alternative approach to do
something similar without an ItemDisambiguationFactory.

Bug: T281721
Change-Id: Ia4249f7662ba17772804123c8c9c7b4c69b4ea14