Remove fallback indicator for variant and mul fallbacks
Instead of hiding the fallback indicator via CSS, completely remove the
HTML for it. (The styles are kept for a little while to support old HTML
in the parser cache.) This also means we can put the no-break space
outside the element again, effectively reverting part of change
I8dada901bf (commit b8c6bc4fb2), if I’m not mistaken.

The removal of the wb-language-fallback-variant class also includes a
minor behavior change. Previously, elements where Wikibase detected both
a transliteration and a variant fallback were tagged with
wb-language-fallback-transliteration and wb-language-fallback-variant;
the latter class then caused the whole element to be hidden. Now, the
element will be shown (and tagged only with -transliteration). I have
not been able to reproduce such a case with any real languages
(LabelsProviderEntityIdHtmlLinkFormatterTest fakes it with
$enGbEnCaTerm), and I suspect it’s not possible – existing
LanguageConverters seem to be able to convert between any languages with
the same base language code (e.g. any zh-*), so the variant case never
applies. If it does apply, then I think the previous behavior was a
mistake, and showing the transliteration indicator should be more
important than hiding the variant indicator.

Bug: T345883
Bug: T291817
Bug: T345881
Bug: T345396
Change-Id: I894ef4df05b6c45a8e5f1a53b46ea73cdff6d276
Depends-On: I4820495f5bed039d2fa410621acafdd1188eb7cb
Depends-On: I615f0b07bea970e81559d12de89eae9915bdeacd