Add a few strict types to help the MediaWiki CodeSniffer
This is mostly motivated by the (currently disabled) MediaWiki
CodeSniffer sniff that requires types to be documented, either with
a PHPDoc comment or a strict type. In these cases here I feel like
a strict type is the better, more sustainable fix. I touched only
places in the code where it's really, really obvious that this is
the correct type.

Change-Id: I89f202af1d05387c8944c83104ba1d7c642e6475