Enable some PHPCS sniffs to avoid common mistakes in PHPDoc blocks
The two FunctionComment and PropertyDocumentation sniffs are extremely
complex by now. They can find many dozen issues. We can enabled and
disabled them individually.

We can relatively easily start making use of this and let the sniffs
find the most obvious mistakes in PHPDoc comments, as demonstrated by
this patch. This still leaves almost all the flexibility that comes
with disabling some of the more annoying rules.

Also: Some of the reasons why we disabled these sniffs a long time ago
are bugs that are fixed by now.

Change-Id: I8952027d89862d8511125dbbd9e131df3274a7ff