EntityId: Remove Serializable interface and methods
This removes the ability to deserialize old serializations, but nothing
seems to rely on that ability: we hard-deprecated unserialize() and
didn’t see any log messages for it in production.

Also remove all the test cases that are “kind of an injection vector”.
__unserialize() still allows injecting some invalid serializations like
that, but it’s not something we want to support, so I have no idea why
we would want to test it, quite frankly.

Bug: T345856
Change-Id: I4850feb0b58ec6367c837f3855af50d34eedd665
Depends-On: I16cf6dfaa4c6271a158c62b4bc55372a7520488e