Split parser cache by desktop/mobile
As long as we use the “mobile view” flag to decide whether to add
certain ResourceLoader modules to the parser output or not, we also have
to split the parser output by this flag, otherwise mobile parser output
may be shown on desktop page views or vice versa.

In production, this should only affect lexemes (and possibly MediaInfo?
no idea), but not items and properties, since for items and properties
the termbox flag already effectively splits the parser cache between
desktop and mobile. And since termbox has been enabled by default (see
change Ib14ae64930, commit d9098551e4), the same would also be true for
most other wikis; however, wikis that disable the mobile termbox might
also have observed the broken styles on items and properties, I believe.

Following some extensive testing and debugging on localhost and
mwdebug2002 (see the comments of T344362, especially T344362#9235624 and
T344362#9250173), I’m pretty confident that this will *not* effectively
clear the parser cache immediately (which would be catastrophic). Old
parser cache contents for desktop and/or mobile will continue to be
valid. However, once a new parser cache entry for a page is saved,
whether for desktop or mobile (e.g. when a mobile page view comes in for
an item, and the parser cache only has a desktop page view so far), then
this will cause the recorded “ParseUsedOptions” to include wbMobile,
whereupon the next request for the other kind of page view will also
regenerate the parser output so that it includes this option. (That is
to say, if an item had only a desktop result in the parser cache, then
the next mobile page view after this change is deployed will cause that
desktop result to be considered invalid as well. Note that this mainly
applies to items and properties, which were already split due to the
termboxVersion; the parser cache for lexemes is additionally affected by
T348872, which will be addressed separately.)

Bug: T344362
Change-Id: Id990974c4980d694f297415b33a22933678b8c57