Enable FunctionComment.SpacingAfterParamName sniff
Trying to align comments vertically can never really work, as
demonstrated by this patch. In most places that once tried to do
this it turned into a chaotic mixture by now. E.g. because someone
added a parameter. Or because a type was changed from e.g. "array"
to "string[]", which is 3 characters longer and messes up all
alignment. A 1-line change as trivial as this suddenly forces us to
touch all other lines as well. Why?

That's why the MediaWiki CodeStyle goes so far and reports the use
of multiple spaces as an error.

Only very few of such places exist in this codebase. Which means
almost all code was already following this rule. I believe that's
argument enough to enable and enforce it. I hope the current owners
of this codebase agree.

Change-Id: Icdd9db9092cb41c9a22e0ae3f239f88f42079c60