jquery.wikibase.wbtooltip: Fix offsets
This fixes the left offset for s* orientation:
This has been corrected for the top offset before,
but the left offset was left out. This incomplete
placement correction logic was introduced in 0f23aca3d31.

Also when setting up tipsy, set the element’s title to the
HTML of the contents. tipsy will use this attribute to
initialize the contents (note that we already set html:
true in line 83, so tipsy will use the attribute as HTML,
not as text), which means it can do the size calculations
correctly. (We’ll still replace the contents later, but at
least tipsy will have had the right information for the
layout.) We need at least the width to be correct so that
the ne gravity will actually move the tooltip far enough to
the left.

Bug: T343998
Change-Id: Ia642fffde449b388fb32960303aa75f6e7e192e2
Co-Authored-By: Lucas Werkmeister <lucas.werkmeister@wikimedia.de>