Revert "Add hook to invalidate cache entries missing TermboxOption"
The hook was only meant to be temporary (T228978#5377451), and I think
it’s no longer needed: every parser cache entry contains *some*
termboxVersion now (either 1 for no / legacy termbox or a higher value
such as 22 for the mobile termbox), so when the termboxEnabled setting
is changed, the parser cache is already split on that and the error from
T228978 shouldn’t happen anymore.

At the same time, this hook handler was causing all parser cache entries
for non-item/property pages to be rejected on mobile, which we
definitely don’t want (see T348872), so we’re quite keen to either fix
it (to accommodate non-item/property pages) or to remove it; if it’s not
neeeded, then removing it seems like the easier option.

This reverts commit bde219a08413c6eec355ed7c7734ed6bbe54474e.

Bug: T348872
Change-Id: I59a0aca0c5d68efaa3619e8b1f2e263285135a5b