Enable FunctionComment.SpacingAfterParamType PHPCS rule
The argument is the same as in Icdd9db9: Trying to align comments
vertically can never really work. In most places that once tried to
do this it turned into something else by now. E.g. because someone
added a parameter. Or because a type was changed from e.g. "array"
to "string[]" which is 3 characters longer and messes up all
alignment. What should be a 1-line change suddenly forces people to
touch other lines for no good reason.

That's why the MediaWiki CodeStyle goes so far and reports the use
of multiple spaces as an error.

Only very few of such places exist in this codebase. Which means
almost all code was already following this rule. That might be
argument enough to enable and enforce it. However, this particular
rule has much less of an impact. The current authors of this
codebase should decide how to proceed.

Change-Id: Ifb02155d9df0975d68772f1b6b767700af897908