Fix and enable FunctionComment.MissingReturn PHPCS rule
It looks like this is worth it. There are not that many issues
found, which implies that almost all of this codebase was already
following this rule.

I tried to apply the most minimal fixes as well as reduce
duplication by utilizing the @inheritDoc feature. It effectively
marks a method as "intentionally undocumented".

Change-Id: I394ad1fc9b918b0b7896bb47d60723587e174f83