Add temporary user support to most API modules
Create a temporary user if appropriate when an entity is modified, and
if a temp user has been created then append the new username to API
responses. This should cover everything that uses MediaWikiEditEntity;
notably, it does *not* cover EntityRedirectCreationInteractor or
ItemMergeInteractor, which use the underlying EntityStore directly
(compare T345064).

Bug: T349130
Bug: T354730
Depends-On: I4fd445df67a63c5e357a3693e7f71519e5ebdd25
Change-Id: I57075e886937ad088a215ff36c22eb7fe3fb2a3c
Co-Authored-By: Cormac Parle <cparle@wikimedia.org>