REST: Isolate flaky e2e test
Since all authorization tests are expected to have the same response, we
have observed the cause of the flakiness (page protection/unprotection
not taking effect immediately) to also cause false positives in the
other tests.
This moves the flaky "protected entity page" test into a separate loop
and to the end of the suite so that the side effects cannot spill over
into other tests.

Change-Id: I27e49ff8e6371f90e1d464cd1d8870c52b30841e