REST: Enable deep linking in Swagger docs
When `deepLinking` is enabled, Swagger UI will automatically update its URL fragment with a deep link to the tag or operation that is expanded. It also enables copying the deep link from the right click context menu.

This will make it easier to create deep links to specific endpoints without having to guess or inspect the page source.

Change-Id: Ib6f796fdf11499cd3f24986e3a675a411e88986a