Adjust styling for sitelink groups to avoid horizontal scrolling
Currently the sitelink groups have a fixed width of 30em, which can
cause horizontal scrolling on mobile and other narrow screens. This
makes the sections as wide as possible with a maximum width of 30em,
which should behave the same as it currently does on wider screens,
while allowing it to be narrower when necessary.

This has been in Wikidata's MediaWiki:Common.css for the past two
months [1] and does not seem to have caused any problems.

[1] https://www.wikidata.org/w/index.php?diff=1970446149

Bug: T330820
Change-Id: I1174447eaba0c462a1aa07a5f1109de269c45fae