Fix broken baserevid comparison
And update the test so that it actually covers the comparison as
intended: change the data so that, were it not for the baserevid, it
would merge cleanly. Previously, it produced an edit conflict later in
the edit which the test then misinterpreted as an edit conflict from the
baserevid comparison.

Change-Id: I4ff3aca1900f187882aaf808fa3f6941a6d6dc64