Move mul onboarding popup below pulsating dot
For the popup itself, this is as simple as changing the 'position' from
'after' to 'below', but we need to make some additional changes to make
the anchor of the popup look right.

There are now three elements: the “container”, which contains the “popup
container”, which contains the actual “pulsating dot”. The “container”
has the click handler and most of the important styles, including a
non-zero width so that it overflows when necessary; the “popup
container” is what we attach the popup to, and has zero width so that
the anchor points to the middle of the dot; the dot itself is styled by
MediaWiki, as before. I tried to reduce this to two elements, but I
couldn’t work out a way to do it: attaching the popup to the outer
container breaks the anchor’s horizontal position due to its width,
while attaching the popup to the dot directly breaks the anchor’s
vertical position.

Bug: T345750
Change-Id: Ida96287d740f40dfc0e2d740691948d50dbea4e6