Change name of mul language when used for terms
The mul language code should continue to be called “multiple languages”
for monolingual text values, but “default values (mul)” for terms
(mainly labels). To accommodate this, add a second method to
LanguageNameLookup – getName() does the same thing as before, while
getNameForTerms() returns the special name for mul in a terms context
using a custom message (and delegates to getName() for all other
language codes).

I think the custom message is currently only reachable via repo code,
but as the code using the message itself is in lib, I think it’s safer
to put the message in lib/i18n/ rather than repo/i18n/.

Bug: T345083
Change-Id: I7dbd94be1313561d4e08cd0a277b30815d939d09