Use LanguageNameUtils::AUTONYMS constant in test, not null
Seeing null there is a bit confusing when looking at the dataProvider,
given that none of the cases seem to include null. Using the constant
LanguageNameUtils::AUTONYMS instead, will make the intention of that
condition much clearer. That this constant is actually null is still
being hinted by the ?string typehint in the signature.

This basically reapplies to change in PS3 from I26281cf904d, see
https://gerrit.wikimedia.org/r/c/mediawiki/extensions/Wikibase/+/977257/2..3

Change-Id: Iee1c5143e92db2adc47bfc9a38e781565e6089a1