Update copy of MUL onboarding popup to "default labels"
This explains the purpose behind introducing mul better. Also, it is a
piece of the puzzle of providing consistent branding around "default
labels".

The link is now no longer part of the message, so that it can be changed
without the message having to be retranslated. During user-testing, it will
point to a temporary location as a hard-coded value.

Bug: T345083
Change-Id: Id4a5b6f5d93efa74270c985dc81fb20ea7dd4c5c