Remove fallback dead code from languageLabelDescriptionAliases.js
getLanguageNameByCode() cannot return a falsy value – it returns the
language code as a last fallback – so the fallback here to the language
code in brackets is pointless and can be removed. (If the language code
itself is falsy, i.e. the empty string, then I don’t think changing it
to `[]` is any better.)

Change-Id: I3288ef0f947cbe23dd35277ffd862bcf85430625